+++ Source/Scripts/Helpers/Invoke-Restart.ps1
@@ -5,8 +5,8 @@ param (
     
 $cmdletsPaths="$PSScriptRoot\..\..\Cmdlets"
 
-. "$cmdletsPaths\Helpers\Write-MyInvocation.ps1"
-Write-MyInvocation -Invocation $MyInvocation
+. "$cmdletsPaths\Helpers\Write-Separator.ps1"
+Write-Separator -Invocation $MyInvocation -Header
 
 . "$cmdletsPaths\Helpers\Invoke-CommandWrap.ps1"
 
@@ -53,3 +53,4 @@ do {
 }while(-not $areAlive)
 
 Write-Host "$Computer is back online"
+Write-Separator -Invocation $MyInvocation -Footer