+++ Source/Scripts/Helpers/Kill-RemoteSessions.ps1
@@ -5,8 +5,8 @@ param (
 
 $cmdletsPaths="$PSScriptRoot\..\..\Cmdlets"
 
-. "$cmdletsPaths\Helpers\Write-MyInvocation.ps1"
-Write-MyInvocation -Invocation $MyInvocation
+. "$cmdletsPaths\Helpers\Write-Separator.ps1"
+Write-Separator -Invocation $MyInvocation -Header
 
 . "$cmdletsPaths\Helpers\Invoke-CommandWrap.ps1"
 
@@ -16,3 +16,4 @@ $killRemoteSessionsScriptBlock={
 }
 
 Invoke-CommandWrap -ComputerName $Computer -ScriptBlock $killRemoteSessionsScriptBlock -BlockName "Kill remote sessions"
+Write-Separator -Invocation $MyInvocation -Footer