+++ Source/Scripts/IIS/New-IISSslBinding.ps1
@@ -4,8 +4,8 @@ param (
 )
 $cmdletsPaths="$PSScriptRoot\..\..\Cmdlets"
 
-. "$cmdletsPaths\Helpers\Write-MyInvocation.ps1"
-Write-MyInvocation -Invocation $MyInvocation
+. "$cmdletsPaths\Helpers\Write-Separator.ps1"
+Write-Separator -Invocation $MyInvocation -Header
 
 . "$cmdletsPaths\Helpers\Invoke-CommandWrap.ps1"
 
@@ -43,3 +43,5 @@ try
 finally
 {
 }
+
+Write-Separator -Invocation $MyInvocation -Footer