+++ Source/Scripts/PackageManagement/Install-NugetPackageProvider.ps1
@@ -5,8 +5,8 @@
 
 $cmdletsPaths="$PSScriptRoot\..\..\Cmdlets"
 
-. "$cmdletsPaths\Helpers\Write-MyInvocation.ps1"
-Write-MyInvocation -Invocation $MyInvocation
+. "$cmdletsPaths\Helpers\Write-Separator.ps1"
+Write-Separator -Invocation $MyInvocation -Header
 
 . "$cmdletsPaths\Helpers\Invoke-CommandWrap.ps1"
 
@@ -26,5 +26,4 @@ catch
     Write-Error $_
 }
 
-
-
+Write-Separator -Invocation $MyInvocation -Footer