+++ Source/Scripts/PowerShellGet/Register-Repository.ps1
@@ -16,8 +16,8 @@
 
 $cmdletsPaths="$PSScriptRoot\..\..\Cmdlets"
 
-. "$cmdletsPaths\Helpers\Write-MyInvocation.ps1"
-Write-MyInvocation -Invocation $MyInvocation
+. "$cmdletsPaths\Helpers\Write-Separator.ps1"
+Write-Separator -Invocation $MyInvocation -Header
 
 . "$cmdletsPaths\Helpers\Invoke-CommandWrap.ps1"
 
@@ -50,5 +50,4 @@ catch
     Write-Error $_
 }
 
-
-
+Write-Separator -Invocation $MyInvocation -Footer